practices/rag-practice/requirements.txt
psycopg2-binary
sentence-transformers
google-generativeai
python-dotenv